backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
pymongo
django-cors-headers

mongoengine
requests